requirements.txt
pytz
requests
python-dotenv==0.20.0
click==8.0.3
dj-database-url==1.0.0

# Django
django
django-cleanup==5.2.0
django-timezone-field 
django-cors-headers
pillow==8.2.0
whitenoise

# Django 3.0 async requirements
gunicorn==20.1.0
uvicorn
uvloop
httptools

# Databases
psycopg2-binary

# Telegram
python-telegram-bot
python-telegram-bot-pagination==0.0.2

